requirements.txt
uvicorn==0.21.1
python-multipart==0.0.6
fastapi==0.95.0
nltk==3.6.7
numpy==1.21.2
Pillow==9.4.0
psutil==5.9.4
pydantic==1.8.2
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cpu
torchvision==0.14.0
